pyproject.toml
[tool.poetry]
name = "pgpcrypto"
version = "0.1.0"
description = ""
authors = ["Nathan Hadzariga <nathan.hadzariga@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-gnupg = "^0.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
boto3 = "^1.34.26"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
